requirements.txt
# Core Data Science
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
imbalanced-learn>=0.11.0

# Hyperparameter Tuning
optuna>=3.4.0

# Deep Learning (CPU-only)
tensorflow-cpu>=2.14.0

# Explainability
shap>=0.43.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Utilities
joblib>=1.3.0
tqdm>=4.66.0
